Move transaction typing into QuillStorageCells
